+++ .sequelizerc
@@ -1,0 +1,7 @@
+const path = require('path');
+
+module.exports = {
+  'models-path': path.resolve('src', 'models'),
+  'migrations-path': path.resolve('src', 'db', 'migrations'),
+  'seeders-path': path.resolve('src', 'db', 'seeders'),
+};
