+++ src/db/dbUtil.ts
@@ -1,8 +1,8 @@
-import db from './index';
+import sequelize from './index';
 
 export async function testAuthenticate() {
   try {
-    await db.authenticate();
+    await sequelize.authenticate();
     console.log('Connection has been established successfully');
   } catch (err) {
     console.error('Unable to connect to db:', err);
