+++ src/db/index.ts
@@ -1,6 +1,6 @@
 import {Sequelize} from 'sequelize';
 
 const connectionUri = process.env.DATABASE_URL || '';
-const db = new Sequelize(connectionUri);
+const sequelize = new Sequelize(connectionUri);
 
-export default db;
+export default sequelize;
